+++ src/components/Icon/MinusIcon/examples/1.default.jsx
@@ -1,14 +1,4 @@
 import React from 'react';
-import createClass from 'create-react-class';
 import { MinusIcon } from '../../../../index';
 
-export default createClass({
-	render() {
-		return (
-			<div>
-				<MinusIcon />
-				<MinusIcon isBadge />
-			</div>
-		);
-	},
-});
+export default () => <MinusIcon />;