+++ src/components/Icon/PlusIcon/examples/1.default.jsx
@@ -1,14 +1,4 @@
 import React from 'react';
-import createClass from 'create-react-class';
 import { PlusIcon } from '../../../../index';
 
-export default createClass({
-	render() {
-		return (
-			<div>
-				<PlusIcon />
-				<PlusIcon isBadge />
-			</div>
-		);
-	},
-});
+export default () => <PlusIcon />;