+++ src/components/Icon/SettingsIcon/examples/1.default.jsx
@@ -1,15 +1,4 @@
 import React from 'react';
-import createClass from 'create-react-class';
 import { SettingsIcon } from '../../../../index';
 
-export default createClass({
-	render() {
-		return (
-			<div>
-				Small: <SettingsIcon />
-				<br />
-				Large: <SettingsIcon presetSize='large' />
-			</div>
-		);
-	},
-});
+export default () => <SettingsIcon />;