+++ src/components/Icon/StopwatchIcon/examples/1.default.jsx
@@ -1,9 +1,4 @@
 import React from 'react';
-import createClass from 'create-react-class';
 import { StopwatchIcon } from '../../../../index';
 
-export default createClass({
-	render() {
-		return <StopwatchIcon />;
-	},
-});
+export default () => <StopwatchIcon />;