+++ src/components/Icon/SuccessIcon/SuccessIcon.jsx
@@ -25,28 +25,16 @@ const SuccessIcon = createClass({
 	},
 
 	render() {
-		const { className, isDisabled, isClickable, ...passThroughs } = this.props;
+		const { className, ...passThroughs } = this.props;
 
 		return (
 			<Icon
 				{...omitProps(passThroughs, SuccessIcon, [], false)}
 				{..._.pick(passThroughs, _.keys(Icon.propTypes))}
-				isClickable={isClickable}
-				isDisabled={isDisabled}
-				className={cx('&', className, isClickable && '&-is-clickable')}
+				className={cx('&', className)}
 			>
-				<circle
-					className={cx('&-background', {
-						'&-background-is-disabled': isDisabled,
-					})}
-					cx='8'
-					cy='8'
-					r='8'
-				/>
-				<path
-					className={cx('&-check')}
-					d='M11.92 5.19l-.22-.23c-.284-.285-.75-.285-1.032 0l-4.13 4.127L4.79 7.474c-.285-.286-.747-.286-1.032 0l-.287.342c-.285.286-.285.746 0 1.033l2.486 2.353c.285.285.748.285 1.033 0l4.93-4.98c.285-.285.285-.748 0-1.034z'
-				/>
+				<circle className={cx('&-background')} cx='8' cy='8' r='7.5' />
+				<path className={cx('&-check')} d='M4.5 8L7 10.5 11.5 6' />
 			</Icon>
 		);
 	},