+++ src/components/Icon/SuccessLightIcon/SuccessLightIcon.jsx
@@ -25,24 +25,16 @@ const SuccessLightIcon = createClass({
 	},
 
 	render() {
-		const { className, isDisabled, isClickable, ...passThroughs } = this.props;
+		const { className, ...passThroughs } = this.props;
 
 		return (
 			<Icon
 				{...passThroughs}
 				{..._.pick(passThroughs, _.keys(Icon.propTypes))}
-				isClickable={isClickable}
-				isDisabled={isDisabled}
-				className={cx('&', className, isClickable && '&-is-clickable')}
+				className={cx('&', className)}
 			>
-				<path
-					className={cx('&-background')}
-					d='M8 15c-3.86 0-7-3.14-7-7s3.14-7 7-7 7 3.14 7 7-3.14 7-7 7z'
-				/>
-				<path
-					className={cx('&-check', { '&-check-is-disabled': isDisabled })}
-					d='M11.92 5.19c.285.284.285.748 0 1.032l-4.932 4.98c-.285.286-.748.286-1.033 0L3.47 8.85c-.286-.287-.286-.748 0-1.034l.287-.343c.285-.286.747-.286 1.033 0l1.747 1.614 4.13-4.127c.284-.285.748-.285 1.033 0l.22.23zM16 8c0 4.418-3.582 8-8 8s-8-3.582-8-8 3.582-8 8-8 8 3.582 8 8zm-1 0c0-3.86-3.14-7-7-7S1 4.14 1 8s3.14 7 7 7 7-3.14 7-7z'
-				/>
+				<path class='st0' d='M4.5 8L7 10.5 11.5 6' />
+				<circle class='st0' cx='8' cy='8' r='7.5' />
 			</Icon>
 		);
 	},