+++ src/components/Icon/SwitchIcon/examples/1.default.jsx
@@ -1,9 +1,4 @@
 import React from 'react';
-import createClass from 'create-react-class';
 import { SwitchIcon } from '../../../../index';
 
-export default createClass({
-	render() {
-		return <SwitchIcon />;
-	},
-});
+export default () => <SwitchIcon />;