+++ src/components/Icon/UnlockedIcon/examples/1.default.jsx
@@ -1,9 +1,4 @@
 import React from 'react';
-import createClass from 'create-react-class';
 import { UnlockedIcon } from '../../../../index';
 
-export default createClass({
-	render() {
-		return <UnlockedIcon />;
-	},
-});
+export default () => <UnlockedIcon />;